+++ packages/block-editor/src/components/block-list/index.js
@@ -8,6 +8,7 @@ import classnames from 'classnames';
  */
 import { AsyncModeProvider, useSelect } from '@wordpress/data';
 import { useRef, createContext, useState } from '@wordpress/element';
+import { useViewportMatch } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -19,15 +20,27 @@ import useInsertionPoint from './insertion-point';
 import BlockPopover from './block-popover';
 import { store as blockEditorStore } from '../../store';
 import { useScrollSelectionIntoView } from '../selection-scroll-into-view';
+import { LayoutProvider, defaultLayout } from './layout';
 
 export const BlockNodes = createContext();
 export const SetBlockNodes = createContext();
 
-export default function BlockList( { className } ) {
+export default function BlockList( { className, __experimentalLayout } ) {
 	const ref = useRef();
 	const [ blockNodes, setBlockNodes ] = useState( {} );
 	const insertionPoint = useInsertionPoint( ref );
 	useScrollSelectionIntoView( ref );
+
+	const isLargeViewport = useViewportMatch( 'medium' );
+	const { isTyping, isOutlineMode, isFocusMode } = useSelect( ( select ) => {
+		const { isTyping: _isTyping, getSettings } = select( blockEditorStore );
+		const { outlineMode, focusMode } = getSettings();
+		return {
+			isTyping: _isTyping(),
+			isOutlineMode: outlineMode,
+			isFocusMode: focusMode,
+		};
+	}, [] );
 
 	return (
 		<BlockNodes.Provider value={ blockNodes }>
@@ -37,11 +50,19 @@ export default function BlockList( { className } ) {
 				ref={ ref }
 				className={ classnames(
 					'block-editor-block-list__layout is-root-container',
-					className
+					className,
+					{
+						'is-typing': isTyping,
+						'is-outline-mode': isOutlineMode,
+						'is-focus-mode': isFocusMode && isLargeViewport,
+					}
 				) }
 			>
 				<SetBlockNodes.Provider value={ setBlockNodes }>
-					<BlockListItems wrapperRef={ ref } />
+					<BlockListItems
+						wrapperRef={ ref }
+						__experimentalLayout={ __experimentalLayout }
+					/>
 				</SetBlockNodes.Provider>
 			</div>
 		</BlockNodes.Provider>
@@ -53,6 +74,7 @@ function Items( {
 	rootClientId,
 	renderAppender,
 	__experimentalAppenderTagName,
+	__experimentalLayout: layout = defaultLayout,
 	wrapperRef,
 } ) {
 	function selector( select ) {
@@ -88,7 +110,7 @@ function Items( {
 	const isAppenderDropTarget = dropTargetIndex === blockClientIds.length;
 
 	return (
-		<>
+		<LayoutProvider value={ layout }>
 			{ blockClientIds.map( ( clientId, index ) => {
 				const isBlockInSelection = hasMultiSelection
 					? multiSelectedBlockClientIds.includes( clientId )
@@ -129,7 +151,7 @@ function Items( {
 						isAppenderDropTarget && orientation === 'horizontal',
 				} ) }
 			/>
-		</>
+		</LayoutProvider>
 	);
 }