+++ packages/block-editor/src/components/block-list/use-block-props/use-block-class-names.js
@@ -7,7 +7,6 @@ import classnames from 'classnames';
  * WordPress dependencies
  */
 import { useSelect } from '@wordpress/data';
-import { useViewportMatch } from '@wordpress/compose';
 import { isReusableBlock, getBlockType } from '@wordpress/blocks';
 
 /**
@@ -23,11 +22,9 @@ import { store as blockEditorStore } from '../../../store';
  * @return {string} The class names.
  */
 export function useBlockClassNames( clientId ) {
-	const isLargeViewport = useViewportMatch( 'medium' );
 	return useSelect(
 		( select ) => {
 			const {
-				isTyping,
 				isBlockBeingDragged,
 				isBlockHighlighted,
 				isBlockSelected,
@@ -38,8 +35,6 @@ export function useBlockClassNames( clientId ) {
 				__experimentalGetActiveBlockIdByBlockNames: getActiveBlockIdByBlockNames,
 			} = select( blockEditorStore );
 			const {
-				focusMode,
-				outlineMode,
 				__experimentalSpotlightEntityBlocks: spotlightEntityBlocks,
 			} = getSettings();
 			const isDragging = isBlockBeingDragged( clientId );
@@ -60,23 +55,12 @@ export function useBlockClassNames( clientId ) {
 				'is-multi-selected': isBlockMultiSelected( clientId ),
 				'is-reusable': isReusableBlock( getBlockType( name ) ),
 				'is-dragging': isDragging,
-				// We only care about this prop when the block is selected
-				// Thus to avoid unnecessary rerenders we avoid updating the prop if
-				// the block is not selected.
-				'is-typing':
-					( isSelected || isAncestorOfSelectedBlock ) && isTyping(),
-				'is-focused':
-					focusMode &&
-					isLargeViewport &&
-					( isSelected || isAncestorOfSelectedBlock ),
-				'is-focus-mode': focusMode && isLargeViewport,
-				'is-outline-mode': outlineMode,
 				'has-child-selected': isAncestorOfSelectedBlock && ! isDragging,
 				'has-active-entity': activeEntityBlockId,
 				// Determine if there is an active entity area to spotlight.
 				'is-active-entity': activeEntityBlockId === clientId,
 			} );
 		},
-		[ clientId, isLargeViewport ]
+		[ clientId ]
 	);
 }