+++ packages/block-editor/src/components/writing-flow/focus-capture.js
@@ -13,7 +13,6 @@ import { useSelect, useDispatch } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { getBlockDOMNode } from '../../utils/dom';
 import { store as blockEditorStore } from '../../store';
 
 /**
@@ -38,6 +37,7 @@ const FocusCapture = forwardRef(
 			isReverse,
 			containerRef,
 			noCapture,
+			lastFocus,
 			hasMultiSelection,
 			multiSelectionContainer,
 		},
@@ -79,20 +79,7 @@ const FocusCapture = forwardRef(
 				return;
 			}
 
-			// If there is a selected block, move focus to the first or last
-			// tabbable element depending on the direction.
-			const wrapper = getBlockDOMNode(
-				selectedClientId,
-				ref.current.ownerDocument
-			);
-
-			if ( isReverse ) {
-				const tabbables = focus.tabbable.find( wrapper );
-				const lastTabbable = last( tabbables ) || wrapper;
-				lastTabbable.focus();
-			} else {
-				wrapper.focus();
-			}
+			lastFocus.current.focus();
 		}
 
 		return (