+++ packages/block-library/src/search/edit.native.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { View, TextInput } from 'react-native';
+import { View } from 'react-native';
 import classnames from 'classnames';
 
 /**
@@ -9,24 +9,51 @@ import classnames from 'classnames';
  */
 import {
 	RichText,
-	BlockControls,
+	PlainText,
 	useBlockProps,
+	InspectorControls,
 } from '@wordpress/block-editor';
-import { ToolbarGroup, ToolbarButton, Button } from '@wordpress/components';
+import {
+	Button,
+	PanelBody,
+	SelectControl,
+	ToggleControl,
+} from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { search } from '@wordpress/icons';
+import { useRef, useEffect, useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { buttonWithIcon, toggleLabel } from './icons';
-import ButtonPositionDropdown from './button-position-dropdown';
 import styles from './style.scss';
 import richTextStyles from './rich-text.scss';
 
+/**
+ * Constants
+ */
 const MIN_BUTTON_WIDTH = 100;
+const BUTTON_OPTIONS = [
+	{ value: 'button-inside', label: __( 'Button inside' ) },
+	{ value: 'button-outside', label: __( 'Button outside' ) },
+	{ value: 'no-button', label: __( 'No button' ) },
+];
 
-export default function SearchEdit( { attributes, setAttributes, className } ) {
+export default function SearchEdit( {
+	onFocus,
+	isSelected,
+	attributes,
+	setAttributes,
+	className,
+} ) {
+	const [ isButtonSelected, setIsButtonSelected ] = useState( false );
+	const [ isLabelSelected, setIsLabelSelected ] = useState( false );
+	const [ isPlaceholderSelected, setIsPlaceholderSelected ] = useState(
+		true
+	);
+
+	const textInputRef = useRef( null );
+
 	const {
 		label,
 		showLabel,
@@ -35,6 +62,20 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 		placeholder,
 		buttonText,
 	} = attributes;
+
+	/*
+	 * Called when the value of isSelected changes. Blurs the PlainText component
+	 * used by the placeholder when this block loses focus.
+	 */
+	useEffect( () => {
+		if ( hasTextInput() && isPlaceholderSelected && ! isSelected ) {
+			textInputRef.current.blur();
+		}
+	}, [ isSelected ] );
+
+	const hasTextInput = () => {
+		return textInputRef && textInputRef.current;
+	};
 
 	const getBlockClassNames = () => {
 		return classnames(
@@ -60,47 +101,57 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 		);
 	};
 
+	const getSelectedButtonPositionLabel = ( option ) => {
+		switch ( option ) {
+			case 'button-inside':
+				return __( 'Inside' );
+			case 'button-outside':
+				return __( 'Outside' );
+			case 'no-button':
+				return __( 'No button' );
+		}
+	};
+
 	const blockProps = useBlockProps( {
 		className: getBlockClassNames(),
 	} );
 
 	const controls = (
-		<BlockControls>
-			<ToolbarGroup>
-				<ToolbarButton
-					title={ __( 'Toggle search label' ) }
-					icon={ toggleLabel }
-					onClick={ () => {
+		<InspectorControls>
+			<PanelBody title={ __( 'Search settings' ) }>
+				<ToggleControl
+					label={ __( 'Hide search heading' ) }
+					checked={ ! showLabel }
+					onChange={ () => {
 						setAttributes( {
 							showLabel: ! showLabel,
 						} );
 					} }
-					isActive={ showLabel }
 				/>
-
-				<ButtonPositionDropdown
-					selectedOption={ buttonPosition }
+				<SelectControl
+					label={ __( 'Button position' ) }
+					value={ getSelectedButtonPositionLabel( buttonPosition ) }
 					onChange={ ( position ) => {
 						setAttributes( {
 							buttonPosition: position,
 						} );
 					} }
+					options={ BUTTON_OPTIONS }
+					hideCancelButton={ true }
 				/>
-
-				{ 'no-button' !== buttonPosition && (
-					<ToolbarButton
-						title={ __( 'Use button with icon' ) }
-						icon={ buttonWithIcon }
-						onClick={ () => {
+				{ buttonPosition !== 'no-button' && (
+					<ToggleControl
+						label={ __( 'Use icon button' ) }
+						checked={ buttonUseIcon }
+						onChange={ () => {
 							setAttributes( {
 								buttonUseIcon: ! buttonUseIcon,
 							} );
 						} }
-						isActive={ buttonUseIcon }
 					/>
 				) }
-			</ToolbarGroup>
-		</BlockControls>
+			</PanelBody>
+		</InspectorControls>
 	);
 
 	const mergeWithBorderStyle = ( style ) => {
@@ -114,7 +165,9 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 				: mergeWithBorderStyle( styles.searchTextInput );
 
 		return (
-			<TextInput
+			<PlainText
+				ref={ textInputRef }
+				isSelected={ isPlaceholderSelected }
 				className="wp-block-search__input"
 				style={ inputStyle }
 				numberOfLines={ 1 }
@@ -124,9 +177,14 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 				placeholder={
 					placeholder ? undefined : __( 'Optional placeholder…' )
 				}
-				onChangeText={ ( newVal ) =>
+				onChange={ ( newVal ) =>
 					setAttributes( { placeholder: newVal } )
 				}
+				onFocus={ () => {
+					setIsPlaceholderSelected( true );
+					onFocus();
+				} }
+				onBlur={ () => setIsPlaceholderSelected( false ) }
 			/>
 		);
 	};
@@ -138,22 +196,32 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 					<Button
 						className="wp-block-search__button"
 						icon={ search }
+						onFocus={ onFocus }
 					/>
 				) }
 
 				{ ! buttonUseIcon && (
 					<RichText
 						className="wp-block-search__button"
+						identifier="text"
+						tagName="p"
 						style={ richTextStyles.searchButton }
 						placeholder={ __( 'Add button text' ) }
 						value={ buttonText }
-						identifier="text"
 						withoutInteractiveFormatting
 						onChange={ ( html ) =>
 							setAttributes( { buttonText: html } )
 						}
 						minWidth={ MIN_BUTTON_WIDTH }
 						textAlign="center"
+						isSelected={ isButtonSelected }
+						__unstableMobileNoFocusOnMount={ ! isSelected }
+						unstableOnFocus={ () => {
+							setIsButtonSelected( true );
+						} }
+						onBlur={ () => {
+							setIsButtonSelected( false );
+						} }
 					/>
 				) }
 			</View>
@@ -167,11 +235,13 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 
 	return (
 		<View { ...blockProps } style={ styles.searchBlockContainer }>
-			{ controls }
+			{ isSelected && controls }
 
 			{ showLabel && (
 				<RichText
 					className="wp-block-search__label"
+					identifier="text"
+					tagName="p"
 					style={ {
 						...styles.searchLabel,
 						...richTextStyles.searchLabel,
@@ -181,6 +251,14 @@ export default function SearchEdit( { attributes, setAttributes, className } ) {
 					withoutInteractiveFormatting
 					value={ label }
 					onChange={ ( html ) => setAttributes( { label: html } ) }
+					isSelected={ isLabelSelected }
+					__unstableMobileNoFocusOnMount={ ! isSelected }
+					unstableOnFocus={ () => {
+						setIsLabelSelected( true );
+					} }
+					onBlur={ () => {
+						setIsLabelSelected( false );
+					} }
 				/>
 			) }